# src/swagger.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "API Pokemon",
    "description": "This API allows you to retrieve pokemons and add, update, delete pokemons\n",
    "version": "1.0.11"
  },
  "tags": [
    {
      "name": "pokemon",
      "description": "Everything about your Pokemon"
    },
    {
      "name": "user",
      "description": "Operations about user"
    }
  ],
  "paths": {
    "/pokemon": {
      "get": {
        "tags": [
          "pokemon"
        ],
        "summary": "Finds all pokemon",
        "description": "Retrieves all pokemon",
        "operationId": "findAllPokemon",
        "responses": {
          "200": {
            "description": "Successful operation"
          }
        }
      },
      "post": {
        "tags": [
          "pokemon"
        ],
        "summary": "Creates a new pokemon",
        "description": "Creates a new pokemon",
        "operationId": "createPokemon",
        "responses": {
          "200": {
            "description": "Successful operation"
          }
        },
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "pokedexNumber": {
                    "type": "number",
                    "example": 1
                  },
                  "name": {
                    "type": "string",
                    "example": "Bulbizarre"
                  },
                  "types": {
                    "type": "array",
                    "items": {
                      "type": "string",
                      "example": "plante"
                    }
                  },
                  "height": {
                    "type": "number",
                    "example": 160
                  },
                  "weight": {
                    "type": "number",
                    "example": 200
                  }
                }
              }
            }
          }
        }
      }
    },
    "/pokemon/{id}": {
      "get": {
        "tags": [
          "pokemon"
        ],
        "summary": "Find pokemon by ID",
        "description": "Retrieves a pokemon by its ID",
        "operationId": "findPokemonById",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "ID of the pokemon to retrieve",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful operation"
          },
          "404": {
            "description": "Pokemon not found"
          }
        }
      },
      "patch": {
        "tags": [
          "pokemon"
        ],
        "summary": "Updates a pokemon",
        "description": "Updates a pokemon",
        "operationId": "updatePokemon",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "ID of the pokemon to update",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful operation"
          },
          "404": {
            "description": "Pokemon not found"
          }
        }
      },
      "delete": {
        "tags": [
          "pokemon"
        ],
        "summary": "Deletes a pokemon",
        "description": "Deletes a pokemon",
        "operationId": "deletePokemon",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "ID of the pokemon to delete",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful operation"
          },
          "404": {
            "description": "Pokemon not found"
          }
        }
      }
    },
    "/user": {
      "post": {
        "tags": [
          "user"
        ],
        "summary": "Create user",
        "description": "Creates a new user",
        "operationId": "createUser",
        "responses": {
          "200": {
            "description": "Successful operation"
          }
        },
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "email": {
                    "type": "string",
                    "example": "user@email.fr"
                  },
                  "password": {
                    "type": "string",
                    "example": "userpass"
                  },
                  "admin": {
                    "type": "boolean",
                    "example": false
                  }
                }
              }
            }
          }
        }
      }
    },
    "/user/login": {
      "post": {
        "tags": [
          "user"
        ],
        "summary": "Logs user into the system",
        "description": "Logs a user into the system",
        "operationId": "loginUser",
        "responses": {
          "200": {
            "description": "Successful operation"
          },
          "400": {
            "description": "Invalid username/password supplied"
          }
        }
      }
    }
  }
}